feat: chashflow is fucking done

+++ src/users/dto/update-user.input.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int } from '@nestjs/graphql';
+import { Field, Float, InputType } from '@nestjs/graphql';
 import { IsOptional, Matches } from 'class-validator';
 
 @InputType()
@@ -20,6 +20,6 @@ export class UpdateUserInput {
   @Field(() => String, { nullable: true })
   cardBandName?: string;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   profitPercent?: number;
 }
